fix: rename action
